# atividade_sprint4/src/main/java/br/com/compasso/partidos/dto/BindFormDTO.java
package br.com.compasso.partidos.dto;

public class BindFormDTO {

	private Long idAssociado;
	private Long idPartido;

	public Long getIdAssociado() {
		return idAssociado;
	}

	public void setIdAssociado(Long idAssociado) {
		this.idAssociado = idAssociado;
	}

	public Long getIdPartido() {
		return idPartido;
	}

	public void setIdPartido(Long idPartido) {
		this.idPartido = idPartido;
	}

}